% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.cell.r
\name{area.cell}
\alias{area.cell}
\title{Get the area of each longitude by latitude cell. Adjust for change in area with latitude.}
\usage{
area.cell(
  lond,
  latd,
  cell.size = rep(1, length(lond)),
  crs.ll = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  crs.en.type = "albers equal area"
)
}
\arguments{
\item{lond}{Vector of longitudes (midpoint of the cell)}

\item{latd}{Vector of latitudes (midpoint of the cell)}

\item{cell.size}{Integer: size of the cell in degrees}

\item{crs.ll}{Character string of the crs for the current lat-lon projections}

\item{crs.en.type}{What type of projection for calculating the area? 'tpeqd', 'mollweide', or 'albers equal area'. Note that the default 'albers equal area' is hard coded for the central pacific}
}
\value{
Either a vector of areas in km2
}
\description{
Get the area of each longitude by latitude cell. Adjust for change in area with latitude.
}
