% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.clockwise.angle.r
\name{calc.clockwise.angle}
\alias{calc.clockwise.angle}
\title{Calculate the clockwise angle between two vectors
Vectors are specified relative to the origin}
\usage{
calc.clockwise.angle(vector1, vector2)
}
\arguments{
\item{vector1}{# vector of length 2. First entry is x and second entry is y}

\item{vector2}{# vector of length 2. First entry is x and second entry is y}
}
\value{
An angle in degrees
}
\description{
Calculate the clockwise angle between two vectors
Vectors are specified relative to the origin
}
