% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaikin.open.r
\name{chaikin.open}
\alias{chaikin.open}
\title{Chaikin corner cutting algorithm for open segments
Used recursively in a loop}
\usage{
chaikin.open(x)
}
\arguments{
\item{x}{Original line segment}
}
\value{
A matrix of points
}
\description{
Chaikin corner cutting algorithm for open segments
Used recursively in a loop
}
