% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{Convert_EN_to_LL_Fn.ndd}
\alias{Convert_EN_to_LL_Fn.ndd}
\title{Function to convert points stored as eastings and northings to lat-lon}
\usage{
Convert_EN_to_LL_Fn.ndd(
  E,
  N,
 
    crs.en = "+proj=tpeqd +lat_1=0 +lon_1=155 +lat_2=0 +lon_2=209 +datum=WGS84 +ellps=WGS84 +units=km +no_defs",
  crs.ll = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{E}{Vector of eastings}

\item{N}{Vector of northings}

\item{crs.en}{Character string of the crs for the E-N projection}

\item{crs.ll}{Character string of the crs for the current lat-lon projections}
}
\value{
Matrix of two columns: Lon and Lat
}
\description{
Function to convert points stored as eastings and northings to lat-lon
}
\keyword{internal}
