% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_LL_to_EastNorth_Fn.ndd.r
\name{Convert_LL_to_EastNorth_Fn.ndd}
\alias{Convert_LL_to_EastNorth_Fn.ndd}
\title{Function to convert points stored as lat-lon to eastings and northings}
\usage{
Convert_LL_to_EastNorth_Fn.ndd(
  Lon,
  Lat,
 
    crs.en = "+proj=tpeqd +lat_1=0 +lon_1=155 +lat_2=0 +lon_2=209 +datum=WGS84 +ellps=WGS84 +units=km +no_defs",
  crs.ll = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{Lon}{Vector of longitudes}

\item{Lat}{Vector of latitudes}

\item{crs.en}{Character string of the crs for the E-N projection}

\item{crs.ll}{Character string of the crs for the current lat-lon projections}
}
\value{
Matrix of two columns: E_km and N_km
}
\description{
Function to convert points stored as lat-lon to eastings and northings
}
