% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Geostat-data.R
\docType{data}
\name{Data_Geostat}
\alias{Data_Geostat}
\title{A simulated data set approximating skipjack biomass in the western and central pacific ocean}
\format{A data frame with 10000 rows and 7 variables
\describe{
  \item{Response_variable}{adult skipjack biomass}
  \item{Year}{timestep}
  \item{Lon}{Longitude}
  \item{Lat}{Latitude}
  \item{Spp}{species code}
  \item{AreaSwept_km2}{area of cell where the data came from in km2}
  \item{Vessel}{filler column incase vessel random effects are desired}
}}
\usage{
Data_Geostat
}
\description{
A simulated data set approximating skipjack biomass in the western and central pacific ocean
}
\keyword{datasets}
