% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro-data.R
\docType{data}
\name{enviro}
\alias{enviro}
\title{A a named-list of length two is required: "formula" is a character string that can be coerced to a formula using \code{stats::as.formula}, and "covariate_data" is a data frame with the following columns - Year, Lon, Lat, covariates...}
\format{
An object of class \code{list} of length 2.
}
\usage{
enviro
}
\description{
A a named-list of length two is required: "formula" is a character string that can be coerced to a formula using \code{stats::as.formula}, and "covariate_data" is a data frame with the following columns - Year, Lon, Lat, covariates...
}
\keyword{datasets}
