% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covariates.ndd.r
\name{make_covariates.ndd}
\alias{make_covariates.ndd}
\title{Is exactly the same as JT FishStatUtils function except it added `colnames(DF_ip) = c(names(sample_data),covariate_names)` to correct column mismatch
This issue has been raised with JT and should be corrected in the next merge of the FishStatUtils dev branch}
\usage{
make_covariates.ndd(
  formula = ~0,
  covariate_data,
  Year_i,
  spatial_list,
  extrapolation_list
)
}
\arguments{
\item{formula}{Formula for the covariates}

\item{covariate_data}{Data frame containing the covariates}

\item{Year_i}{Vector of year of every observation}

\item{spatial_list}{spatial_list produced by \code{make_spatial_info.ndd}}

\item{extrapolation_list}{extrapolation_list produced by \code{make_extrapolation_info.ndd}}
}
\value{
returns a tagged list used in other functions
}
\description{
Is exactly the same as JT FishStatUtils function except it added `colnames(DF_ip) = c(names(sample_data),covariate_names)` to correct column mismatch
This issue has been raised with JT and should be corrected in the next merge of the FishStatUtils dev branch
}
