% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_extrapolation_info.ndd.r
\name{make_extrapolation_info.ndd}
\alias{make_extrapolation_info.ndd}
\title{modified from Jim Thorson's FishStatUtils::Prepare_User_Extrapolation_Data_Fn.R
uses my \code{Convert_EN_to_LL_Fn.ndd} and \code{Convert_LL_to_EastNorth_Fn.ndd} functions
also modifies \code{strata.limits} to match \code{strata.sp} if it is present}
\usage{
make_extrapolation_info.ndd(
  Region,
  strata.limits,
  input.grid,
  crs.en,
  crs.ll,
  strata.sp
)
}
\arguments{
\item{Region}{user}

\item{strata.limits}{data frame with strata limits}

\item{input.grid}{user created extrapolation grid}

\item{crs.en}{Character string of the crs for the E-N projection}

\item{crs.ll}{Character string of the crs for the current lat-lon projections}

\item{strata.sp}{[Optional] If present, a shapefile containing the strata boundaries to calculate the indicies for}
}
\value{
returns a tagged list used in other functions
\describe{
  \item{a_el}{The area associated with each extrapolation grid cell (rows) and strata (columns)}
  \item{Data_Extrap}{A data frame describing the extrapolation grid with columns: Lon, Lat, Area_km2, Include, E_km, N_km. Include appears to be set to 1 for all areas? or set it to 0 for grid cells where you don't want an extrapolation??}
  \item{Area_km2_x}{the area associated with each row of Data_Extrap, in units square-kilometers}
  \item{zone}{}
  \item{flip_around_dateline}{}
}
}
\description{
modified from Jim Thorson's FishStatUtils::Prepare_User_Extrapolation_Data_Fn.R
uses my \code{Convert_EN_to_LL_Fn.ndd} and \code{Convert_LL_to_EastNorth_Fn.ndd} functions
also modifies \code{strata.limits} to match \code{strata.sp} if it is present
}
