% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacific.coast-data.R
\docType{data}
\name{pacific.coast}
\alias{pacific.coast}
\title{A shapefile of coastline centered on the pacific ocean}
\format{
An object of class SpatialPolygons
\describe{
  \item{x min}{75}
  \item{x max}{310}
  \item{y min}{-68.91981}
  \item{y max}{70}
  \item{proj4string}{+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0}
}
}
\source{
<https://www.ngdc.noaa.gov/mgg/shorelines/>
}
\usage{
pacific.coast
}
\description{
A shapefile of coastline centered on the pacific ocean
}
\keyword{datasets}
