% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dharma.resid.r
\name{plot.dharma.resid}
\alias{plot.dharma.resid}
\title{Plot DHARMa style residuals for VAST model.}
\usage{
plot.dharma.resid(
  vast_output,
  residuals_summary,
  n.samp = 999999,
  seed = 123,
  plot.type = "all",
  palette.cols = c("royalblue3", "deepskyblue1", "gold", "orange1", "indianred1",
    "firebrick2", "#AC2020"),
  smooth.span = 0.1,
  ts.start = 2004,
  ts.step = 0.25,
  factor.grouping = 2,
  save.dir,
  save.name
)
}
\arguments{
\item{vast_output}{Output from a call to FishStatsUtils::fit_model}

\item{residuals_summary}{Output from summary() applied to output from FishStatsUtils::fit_model when argument 'what' = "residuals"}

\item{n.samp}{Integer giving the number of records to plot the QQ relationship. Default is to use everyting.}

\item{seed}{The random seed for the subsampling}

\item{plot.type}{The type of metric to plot 'all', 'time', or 'knot', if you want to plot QQ by factor.
\describe{
  \item{'all'}{Observed nominal cpue by knot and timeblock}
  \item{'time'}{Predicted density cpue by knot and timeblock}
  \item{'knot'}{Predicted encounter rate by knot and timeblock}
}}

\item{palette.cols}{A color palette to use for plotting}

\item{smooth.span}{Value between 0 and 1. Degree of smoothing to use for the QQ plot}

\item{ts.start}{Integer giving the first year of the model}

\item{ts.step}{The timestep of the model 1 for year and 0.25 for quarter}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot DHARMa style residuals for VAST model.
}
