% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knots.r
\name{plot.knots}
\alias{plot.knots}
\title{Given fit_vast output, plot the knots and regional structure}
\usage{
plot.knots(vast_output, coast.shp, region.shp.list, save.dir)
}
\arguments{
\item{vast_output}{Output from a call to FishStatsUtils::fit_model}

\item{coast.shp}{supply a coast shapefile, if missing this defaults to the coast shapefile stored in the package}

\item{region.shp.list}{supply a shapefile with the regional structure, if missing this defaults to the WCPO 10N shapefile stored in the package}

\item{save.dir}{directory with the location of where to save the plot}
}
\description{
Given fit_vast output, plot the knots and regional structure
}
