% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qq.wrap.r
\name{plot.qq.wrap}
\alias{plot.qq.wrap}
\title{Plot Q-Q plot for aggregate model fit. Currently only available for delta-lognormal model.}
\usage{
plot.qq.wrap(vast_output, error.structure = "dln", nsims = 1000, save.dir)
}
\arguments{
\item{vast_output}{Output from a call to FishStatsUtils::fit_model}

\item{error.structure}{Character string denoting the error structure used to fit the model. This is needed to calculate the correct quantiles for the residuals.}

\item{nsims}{The number of random draws to use when calculating the quantiles. Default is 1000.}

\item{save.dir}{Path to the directory where the outputs will be saved}
}
\description{
Plot Q-Q plot for aggregate model fit. Currently only available for delta-lognormal model.
}
