% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vast.enviro.response.r
\name{plot.vast.enviro.response}
\alias{plot.vast.enviro.response}
\title{Plot estimated catchability effects. Currently assumes only one set of interactions and no splines on catchability.}
\usage{
plot.vast.enviro.response(
  vast.output,
  enviro.formula,
  report.enviro,
  save.dir,
  save.name
)
}
\arguments{
\item{vast.output}{The output from a fit.vast function call}

\item{enviro.formula}{The formula for the environmental covariates (abundance) that was passed to fit.vast as a part of the enviro argument}

\item{report.enviro}{Named list ('mean', 'sd', 'min', and 'max') where each element of the list is a vector of length covariates in enviro.formula containing the appropriate metric. This is assuming that the environmental covariate was scaled to mean 0 and sd 1 prior to passing to fit.vast.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot estimated catchability effects. Currently assumes only one set of interactions and no splines on catchability.
}
