% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vast.index.r
\name{plot.vast.index}
\alias{plot.vast.index}
\title{Plot index from fit.vast model fit.}
\usage{
plot.vast.index(
  species,
  model.name,
  vast.frq.index,
  which.idx = "idx.std",
  region.names = paste0("R", 1:9),
  smooth.span = 0.1,
  pt.alpha = 0.1,
  plot.uncertainty = FALSE,
  scale.type = "fixed",
  color.palette = c("royalblue3", "deepskyblue1", "gold", "orange1", "indianred1",
    "firebrick2", "#AC2020"),
  save.dir,
  save.name
)
}
\arguments{
\item{species}{String denoting the species of the analysis}

\item{model.name}{What is the model name? Character string.}

\item{vast.frq.index}{Output of vast.frq.index.}

\item{which.idx}{Which index to plot either "idx.std" or "idx.frq"}

\item{region.names}{What should these regions be called? A character vector of names please}

\item{smooth.span}{Degree of smoothing used. Value between 0 and 1, where closer to 1 is more smooth.}

\item{pt.alpha}{alpha transparency of quarterly points or uncertainty band.}

\item{plot.uncertainty}{TRUE or FALSE: Plot the uncertainty.}

\item{scale.type}{Either have yaxis fixed by panel ("fixed") or free ("free_y")}

\item{color.palette}{Color palette used to differentiate betweeen models. Give exact number of colors otherwise a colorRamp will be used.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot index from fit.vast model fit.
}
