% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vast.influ.r
\name{plot.vast.influ}
\alias{plot.vast.influ}
\title{Influence plot for fitted model from fit.vast. Currently assumes only one set of interactions and no splines on catchability.}
\usage{
plot.vast.influ(
  vast.output,
  model.start.year = 1952,
  coef.names = c("flg.grp", "syn.hbf.imb"),
  level.names = c("Flag group", "HBF (pred)"),
  error.structure = "dln",
  pt.alpha = 0.2,
  smooth.span = 0.1,
  save.dir,
  save.name
)
}
\arguments{
\item{vast.output}{The output from a fit.vast function call}

\item{model.start.year}{An integer denoting the first year of data used in the model}

\item{coef.names}{Character vector of the factors used to create Q_ik}

\item{level.names}{Character vector of how to name the factors in the plot}

\item{error.structure}{Character string denoting the error structure used to fit the model. This is needed to backtransform the parameter estimates.}

\item{pt.alpha}{alpha transparency of quarterly points or uncertainty band.}

\item{smooth.span}{Degree of smoothing used. Value between 0 and 1, where closer to 1 is more smooth.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Influence plot for fitted model from fit.vast. Currently assumes only one set of interactions and no splines on catchability.
}
