% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vast.q.estimates.r
\name{plot.vast.q.estimates}
\alias{plot.vast.q.estimates}
\title{Plot estimated catchability effects. Currently assumes only one set of interactions and no splines on catchability.}
\usage{
plot.vast.q.estimates(
  vast.output,
  coef.names = c("flg.grp", "syn.hbf.imb"),
  level.names = c("Flag group", "HBF (pred)"),
  error.structure = "dln",
  save.dir,
  save.name
)
}
\arguments{
\item{vast.output}{The output from a fit.vast function call}

\item{coef.names}{Character vector of the factors used to create Q_ik}

\item{level.names}{Character vector of how to name the factors in the plot}

\item{error.structure}{Character string denoting the error structure used to fit the model. This is needed to backtransform the parameter estimates.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot estimated catchability effects. Currently assumes only one set of interactions and no splines on catchability.
}
