% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vast.space.time.r
\name{plot.vast.space.time}
\alias{plot.vast.space.time}
\title{Plot vast output by spatial knots and time block.}
\usage{
plot.vast.space.time(
  vast_output,
  residuals_summary,
  coast.shp,
  region.shp.list,
  species = "SWO",
  plot.type = "N",
  model.start.year = 2004.25,
  model.int = 0.25,
  t.block = 2,
  plot.obs.grid = FALSE,
  x.extent = c(140, 250),
  y.extent = c(-50, 0),
  scale.trans = "identity",
  viridis.pal = "D",
  save.dir,
  save.name
)
}
\arguments{
\item{vast_output}{Output from a call to FishStatsUtils::fit_model}

\item{residuals_summary}{Output from summary() applied to output from FishStatsUtils::fit_model when argument 'what' = "residuals". Used to get PIT residuals}

\item{coast.shp}{supply a coast shapefile, if missing this defaults to the coast shapefile stored in the package}

\item{region.shp.list}{supply a shapefile with the regional structure, if missing this defaults to the SWPO shapefile stored in the package}

\item{species}{String denoting the species of the analysis}

\item{plot.type}{The type of metric to plot 'pred.cpue', 'bin', 'pos', 'resid', 'bin.resid', 'pos.resid', 'N', 'Omega1', 'Omega2','Epsilon1', or 'Epsilon2'
\describe{
  \item{'obs.cpue'}{Observed nominal cpue by knot and timeblock}
  \item{'pred.cpue'}{Predicted density cpue by knot and timeblock}
  \item{'bin'}{Predicted encounter rate by knot and timeblock}
  \item{'pos'}{Predicted positive catch-rate by knot and timeblock}
  \item{'resid'}{Standardized residual by knot and timeblock}
  \item{'bin.resid'}{Standardized residual for encounter rate cpue by knot and timeblock}
  \item{'pos.resid'}{Standardized residual for positive catch-rate by knot and timeblock}
  \item{'N'}{Number of observations by knot and timeblock}
     \item{'Omega1'}{Spatial random effect for encounter rate}
  \item{'Omega2'}{Spatial random effect for positive catch-rate}
     \item{'Epsilon1'}{Spatiotemporal random effect for encounter rate}
  \item{'Epsilon2'}{Spatiotemporal random effect for positive catch-rate}
}}

\item{model.start.year}{A numeric denoting the first year (or year-quarter) of data used in the model}

\item{model.int}{A numeric denoting the interval used for the temporal sequence}

\item{t.block}{An integer denoting the number of years to aggregate together in time blocks}

\item{plot.obs.grid}{TRUE or FALSE. Plot the observations. Default is FALSE as otherwise the plot looks gross}

\item{x.extent}{Specify the xlim of the plot as c(xmin,xmax) in degrees 0 to 360}

\item{y.extent}{Specify the ylim of the plot as c(ymin,ymax) in degrees -90 to 90}

\item{scale.trans}{A character string denoting the type of ggplot2 scale transformation to use for the color scale. Built-in transformations include "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time".}

\item{viridis.pal}{A character string indicating the colormap option to use. Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E")}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Plot vast output by spatial knots and time block.
}
