% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.all-data.R
\docType{data}
\name{po.all}
\alias{po.all}
\title{A SpatialPolygons object containing the approximate boundaries of the Pacific basin}
\format{
An object of class \code{SpatialPolygons} of length 1.
}
\usage{
po.all
}
\description{
A SpatialPolygons object containing the approximate boundaries of the Pacific basin
}
\keyword{datasets}
