% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.hull.sp.r
\name{smooth.hull.sp}
\alias{smooth.hull.sp}
\title{Calculate a smooth hull around a set of points}
\usage{
smooth.hull.sp(pts, crs.ll, buffer.ll, vertices = 100, k = 3)
}
\arguments{
\item{pts}{List/Data Frame with two entries: one for x and one for y}

\item{crs.ll}{Character string of the crs for the current lat-lon projections}

\item{buffer.ll}{Size of buffer in degrees of lat/lon}

\item{vertices}{Number of vertices in buffer}

\item{k}{Degree of smoothness of buffer, bigger number more smooth (?)}
}
\value{
A spatial polygon object
}
\description{
Calculate a smooth hull around a set of points
}
