% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swpo.swo.shp-data.R
\docType{data}
\name{swpo.swo.shp}
\alias{swpo.swo.shp}
\title{A SpatialPolygons object containing the boundaries of the model regions of the regional structure for the 2021 southwest Pacific Ocean (SWPO) swordfish (SWO) stock assessment}
\format{
An object of class \code{SpatialPolygons} of length 3.
}
\usage{
swpo.swo.shp
}
\description{
A SpatialPolygons object containing the boundaries of the model regions of the regional structure for the 2021 southwest Pacific Ocean (SWPO) swordfish (SWO) stock assessment
}
\keyword{datasets}
