% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swpo.swo.sub.shp-data.R
\docType{data}
\name{swpo.swo.sub.shp}
\alias{swpo.swo.sub.shp}
\title{A SpatialPolygons object containing the boundaries of the model sub-regions of the fisheries structure for the 2021 southwest Pacific Ocean (SWPO) swordfish (SWO) stock assessment}
\format{
An object of class \code{SpatialPolygons} of length 7.
}
\usage{
swpo.swo.sub.shp
}
\description{
A SpatialPolygons object containing the boundaries of the model sub-regions of the fisheries structure for the 2021 southwest Pacific Ocean (SWPO) swordfish (SWO) stock assessment
}
\keyword{datasets}
