% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast.frq.index.r
\name{vast.frq.index}
\alias{vast.frq.index}
\title{Format the estimated index for use in MFCL}
\usage{
vast.frq.index(
  vast_output,
  agg.years = 1980:1990,
  ts.vec = seq(from = 1952, to = 2018.75, by = 0.25),
  region.idx = 3:11,
  region.names = paste0("R", 1:length(region.idx)),
  mean.cv = 0.2,
  missing = 0.05,
  save.dir,
  save.name
)
}
\arguments{
\item{vast_output}{Output from a call to fit.vast where slim output is FALSE}

\item{agg.years}{Years used to define the average period when rescaling}

\item{ts.vec}{a vector denoting the year-quarter of each ts for the fit.vast model output}

\item{region.idx}{The fit.vast model can calculate the abundance trends over many regions, some of which are superfluous. Specify the column index for the regions that you care about.}

\item{region.names}{What should these regions be called? A character vector of names please}

\item{mean.cv}{When rescaling the CV for MFCL, set the mean cv to rescale to.}

\item{missing}{If the cv is missing for some reason (a hold over from the conventional delta-glm) set the penalty weight to this value. The smaller the value, the smaller the impact of this data point on the likelihood in MFCL.}

\item{save.dir}{Path to the directory where the outputs will be saved}

\item{save.name}{Name stem for the output, useful when saving many model outputs in the same directory}
}
\description{
Format the estimated index for use in MFCL
}
