% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.knots.r
\name{plot.knots}
\alias{plot.knots}
\title{Given fit.vast output, plot the knots and regional structure}
\usage{
plot.knots(vast.output, coast.shp, region.shp.list, save.dir)
}
\arguments{
\item{vast.output}{Output from the fit.vast function where slim output is FALSE}

\item{coast.shp}{supply a coast shapefile, if missing this defaults to the pacific coast shapefile stored in the package}

\item{region.shp.list}{supply a shapefile with the regional structure, if missing this defaults to the WCPO skipjack shapefile stored in the package}

\item{save.dir}{directory with the location of where to save the plot}
}
\description{
Given fit.vast output, plot the knots and regional structure
}
